Déplacement composant CardHorizontal dans actions de comm
(car spécifique à cette page)
